pyproject.toml
[project]
name = "boiler-plate-saas"
version = "0.1.0"
description = "Hexagonal Architecture SaaS Boilerplate"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # ORM + async drivers
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",          # PostgreSQL async driver (production)
    "aiosqlite>=0.20.0",        # SQLite async driver (dev / tests)
    # Config
    "pydantic-settings>=2.0.0",
    "pydantic[email]>=2.0.0",
    # Security
    "bcrypt>=4.0.0",
    "alembic>=1.18.4",
    "pyjwt>=2.12.1",
    "structlog>=25.5.0",
    "slowapi>=0.1.9",
    "httpx>=0.27",              # HTTP client for Resend email API
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",              # async test client for FastAPI
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["alembic/"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "B008",   # do not perform function calls in default arguments (FastAPI Depends)
    "UP046",  # Generic[T] syntax — keep for SQLAlchemy/dataclass compatibility
    "E501",   # line too long — handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]
